feat: adding the JWT auth. Adjusting the README. Fixed package.json
This change will add the JWT authorization parts to the udacity-restapi project as described in the lectures.
I've fixed an issue concerning the deployment of the javascript sources in the package.json.
Finally I've adjusted the README.md to explain which steps are necessary to deploy the
project on the aws.